csv-processor-function/requirements.txt
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-pubsub==2.31.1
functions-framework==3.4.0